[test] Fix integer overflow in benchmark size computations
Cast first int operand to size_t in four buffer-size multiplications inside
benchmark_x86_backend.cpp so the entire product is computed in 64-bit arithmetic,
preventing potential int overflow before widening to size_t on large parameter
configurations exercised by the KCache and FP16 attention benchmark fixtures.

+++ test/unittest/benchmark_x86_backend.cpp
@@ -804,9 +804,9 @@ TEST_P(Bench_KCache, compute_kcaches) {
 
   auto in_f32 = generate_random_vector<float>(total_heads * head_dim);
   auto kcache_f32 =
-    generate_random_vector<float>(num_cache_head * seq_len * head_dim);
+    generate_random_vector<float>(static_cast<size_t>(num_cache_head) * seq_len * head_dim);
   auto kcache_u16 = convert_f32_to_f16_u16(kcache_f32);
-  std::vector<float> output(total_heads * seq_len, 0.0f);
+  std::vector<float> output(static_cast<size_t>(total_heads) * seq_len, 0.0f);
 
   auto stats = bench::measure(
     [&]() {
@@ -844,9 +844,9 @@ TEST_P(Bench_Attention, compute_fp16vcache) {
 
   auto in_f32 = generate_random_vector<float>(total_heads * attention_rows);
   auto vcache_f32 =
-    generate_random_vector<float>(num_cache_head * (int)window_size * head_dim);
+    generate_random_vector<float>(static_cast<size_t>(num_cache_head) * (size_t)window_size * head_dim);
   auto vcache_u16 = convert_f32_to_f16_u16(vcache_f32);
-  std::vector<float> output(total_heads * head_dim, 0.0f);
+  std::vector<float> output(static_cast<size_t>(total_heads) * head_dim, 0.0f);
 
   auto stats = bench::measure(
     [&]() {
